+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Ideas/Upcoming]
 
 ## [Unreleased]
+
+## [1.2.0] - 2022-05-29
 ### Changed
-* Replaced KeyboardWatcher by InputWatcher
+* Replaced KeyboardWatcher by InputWatcher with the KeyboardInput.
+### Added
+* Added MouseInput to the InputWatcher.
 
 ## [1.1.3] - 2022-05-14
 ### Added